mcux: Update LPC54114, K64F, and KE1xF to MCUXpresso SDK 2.6.0
Updates the MCUXpresso SDK to version 2.6.0 (2019-06-14) for LPC54114,
K64F, and KE1xF SoCs. This is the latest version available for these
SoCs.

Origin: MCUXpresso SDK
License: BSD 3-Clause
URL: mcux.nxp.com
Maintained-by: External

+++ mcux/drivers/kinetis/fsl_pmc.c
@@ -26,7 +26,7 @@ void PMC_GetParam(PMC_Type *base, pmc_param_t *param)
     uint32_t reg = base->PARAM;
     ;
     param->vlpoEnable = (bool)(reg & PMC_PARAM_VLPOE_MASK);
-    param->hvdEnable = (bool)(reg & PMC_PARAM_HVDE_MASK);
+    param->hvdEnable  = (bool)(reg & PMC_PARAM_HVDE_MASK);
 }
 #endif /* FSL_FEATURE_PMC_HAS_PARAM */
 
@@ -41,14 +41,14 @@ void PMC_GetParam(PMC_Type *base, pmc_param_t *param)
  */
 void PMC_ConfigureLowVoltDetect(PMC_Type *base, const pmc_low_volt_detect_config_t *config)
 {
-    base->LVDSC1 = (0U |
+    base->LVDSC1 = (uint8_t)((0U |
 #if (defined(FSL_FEATURE_PMC_HAS_LVDV) && FSL_FEATURE_PMC_HAS_LVDV)
-                    ((uint32_t)config->voltSelect << PMC_LVDSC1_LVDV_SHIFT) |
+                              ((uint32_t)config->voltSelect << PMC_LVDSC1_LVDV_SHIFT) |
 #endif
-                    ((uint32_t)config->enableInt << PMC_LVDSC1_LVDIE_SHIFT) |
-                    ((uint32_t)config->enableReset << PMC_LVDSC1_LVDRE_SHIFT)
-                    /* Clear the Low Voltage Detect Flag with previouse power detect setting */
-                    | PMC_LVDSC1_LVDACK_MASK);
+                              ((uint32_t)config->enableInt << PMC_LVDSC1_LVDIE_SHIFT) |
+                              ((uint32_t)config->enableReset << PMC_LVDSC1_LVDRE_SHIFT)
+                              /* Clear the Low Voltage Detect Flag with previous power detect setting */
+                              | PMC_LVDSC1_LVDACK_MASK));
 }
 
 /*!
@@ -62,13 +62,13 @@ void PMC_ConfigureLowVoltDetect(PMC_Type *base, const pmc_low_volt_detect_config
  */
 void PMC_ConfigureLowVoltWarning(PMC_Type *base, const pmc_low_volt_warning_config_t *config)
 {
-    base->LVDSC2 = (0U |
+    base->LVDSC2 = (uint8_t)((0U |
 #if (defined(FSL_FEATURE_PMC_HAS_LVWV) && FSL_FEATURE_PMC_HAS_LVWV)
-                    ((uint32_t)config->voltSelect << PMC_LVDSC2_LVWV_SHIFT) |
+                              ((uint32_t)config->voltSelect << PMC_LVDSC2_LVWV_SHIFT) |
 #endif
-                    ((uint32_t)config->enableInt << PMC_LVDSC2_LVWIE_SHIFT)
-                    /* Clear the Low Voltage Warning Flag with previouse power detect setting */
-                    | PMC_LVDSC2_LVWACK_MASK);
+                              ((uint32_t)config->enableInt << PMC_LVDSC2_LVWIE_SHIFT)
+                              /* Clear the Low Voltage Warning Flag with previous power detect setting */
+                              | PMC_LVDSC2_LVWACK_MASK));
 }
 
 #if (defined(FSL_FEATURE_PMC_HAS_HVDSC1) && FSL_FEATURE_PMC_HAS_HVDSC1)
@@ -86,7 +86,7 @@ void PMC_ConfigureHighVoltDetect(PMC_Type *base, const pmc_high_volt_detect_conf
     base->HVDSC1 = (((uint32_t)config->voltSelect << PMC_HVDSC1_HVDV_SHIFT) |
                     ((uint32_t)config->enableInt << PMC_HVDSC1_HVDIE_SHIFT) |
                     ((uint32_t)config->enableReset << PMC_HVDSC1_HVDRE_SHIFT)
-                    /* Clear the High Voltage Detect Flag with previouse power detect setting */
+                    /* Clear the High Voltage Detect Flag with previous power detect setting */
                     | PMC_HVDSC1_HVDACK_MASK);
 }
 #endif /* FSL_FEATURE_PMC_HAS_HVDSC1 */
@@ -105,16 +105,16 @@ void PMC_ConfigureHighVoltDetect(PMC_Type *base, const pmc_high_volt_detect_conf
  */
 void PMC_ConfigureBandgapBuffer(PMC_Type *base, const pmc_bandgap_buffer_config_t *config)
 {
-    base->REGSC = (0U
+    base->REGSC = (uint8_t)(0U
 #if (defined(FSL_FEATURE_PMC_HAS_BGBE) && FSL_FEATURE_PMC_HAS_BGBE)
-                   | ((uint32_t)config->enable << PMC_REGSC_BGBE_SHIFT)
+                            | ((uint32_t)config->enable << PMC_REGSC_BGBE_SHIFT)
 #endif /* FSL_FEATURE_PMC_HAS_BGBE */
 #if (defined(FSL_FEATURE_PMC_HAS_BGEN) && FSL_FEATURE_PMC_HAS_BGEN)
-                   | (((uint32_t)config->enableInLowPowerMode << PMC_REGSC_BGEN_SHIFT))
+                            | (((uint32_t)config->enableInLowPowerMode << PMC_REGSC_BGEN_SHIFT))
 #endif /* FSL_FEATURE_PMC_HAS_BGEN */
 #if (defined(FSL_FEATURE_PMC_HAS_BGBDS) && FSL_FEATURE_PMC_HAS_BGBDS)
-                   | ((uint32_t)config->drive << PMC_REGSC_BGBDS_SHIFT)
+                            | ((uint32_t)config->drive << PMC_REGSC_BGBDS_SHIFT)
 #endif /* FSL_FEATURE_PMC_HAS_BGBDS */
-                       );
+    );
 }
 #endif
